backend/generate_doc_summary_service/requirements.txt
google-cloud-firestore==2.16.0
firebase-admin==6.5.0
functions-framework==3.5.0
google-cloud-storage==2.16.0
google-cloud-secret-manager==2.19.0
google-generativeai==0.7.1
pypdf==4.2.0